add: # 添加用户缓存

+++ src/main/java/com/yp/gameframwrok/engine/message/MessageEvent.java
@@ -7,7 +7,7 @@ import lombok.Data;
 public class MessageEvent {
     private ByteString data;
 
-    private long playerId;
+    private int playerId;
 
     private int mainType;
