+++ src/main/java/com/yp/gameframwrok/game/logic/dn/handler/DnJoinHandler.java
@@ -14,7 +14,7 @@ import org.springframework.stereotype.Component;
 public class DnJoinHandler implements IMessageHandler {
 
     @Override
-    public void handle(ByteString data, long userId) {
+    public void handle(ByteString data, int userId) {
         log.info("ZjhJoinHandler handle data: {}", data);
 
     }
