+++ src/main/java/com/yp/gameframwrok/game/logic/zjh/handler/OperationHandler.java
@@ -14,7 +14,7 @@ import org.springframework.stereotype.Component;
 public class OperationHandler implements IMessageHandler {
 
     @Override
-    public void handle(ByteString data, long userId) {
+    public void handle(ByteString data, int userId) {
         log.info("ZjhOperationHandler handle data: {}", data);
     }
 }
